Renamed package name to "abst"
Still need sth more descriptive
