+++ kumo-cli/src/main/java/com/kennycason/kumo/cli/KumoCli.java
@@ -15,8 +15,8 @@ import com.kennycason.kumo.font.scale.FontScalar;
 import com.kennycason.kumo.font.scale.LinearFontScalar;
 import com.kennycason.kumo.font.scale.LogFontScalar;
 import com.kennycason.kumo.font.scale.SqrtFontScalar;
-import com.kennycason.kumo.interfaces.FontAbst;
-import com.kennycason.kumo.interfaces.InstanceCreator;
+import com.kennycason.kumo.abst.FontAbst;
+import com.kennycason.kumo.abst.InstanceCreator;
 import com.kennycason.kumo.nlp.FrequencyAnalyzer;
 import com.kennycason.kumo.nlp.normalize.*;
 import com.kennycason.kumo.nlp.tokenizer.WhiteSpaceWordTokenizer;
