+++ kumo-core/src/main/java/com/kennycason/kumo/WordCloud.java
@@ -13,7 +13,7 @@ import com.kennycason.kumo.font.scale.LinearFontScalar;
 import com.kennycason.kumo.image.AngleGenerator;
 import com.kennycason.kumo.image.CollisionRaster;
 import com.kennycason.kumo.image.ImageRotation;
-import com.kennycason.kumo.interfaces.*;
+import com.kennycason.kumo.abst.*;
 import com.kennycason.kumo.padding.Padder;
 import com.kennycason.kumo.padding.RectanglePadder;
 import com.kennycason.kumo.padding.WordPixelPadder;
@@ -25,7 +25,6 @@ import com.kennycason.kumo.wordstart.WordStartStrategy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.imageio.ImageIO;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
