+++ kumo-core/src/main/java/com/kennycason/kumo/abst/FontAbst.java
@@ -1,4 +1,4 @@
-package com.kennycason.kumo.interfaces;
+package com.kennycason.kumo.abst;
 
 public abstract class FontAbst<IMPL> implements VaryingImpl<IMPL>{
