+++ kumo-core/src/main/java/com/kennycason/kumo/abst/InstanceCreator.java
@@ -1,4 +1,4 @@
-package com.kennycason.kumo.interfaces;
+package com.kennycason.kumo.abst;
 
 import java.io.InputStream;
