+++ kumo-core/src/main/java/com/kennycason/kumo/abst/VaryingImpl.java
@@ -1,4 +1,4 @@
-package com.kennycason.kumo.interfaces;
+package com.kennycason.kumo.abst;
 
 public interface VaryingImpl<IMPL> {
     IMPL getActual();
