+++ kumo-core/src/main/java/com/kennycason/kumo/awtimpl/DimensionImpl.java
@@ -1,6 +1,6 @@
 package com.kennycason.kumo.awtimpl;
 
-import com.kennycason.kumo.interfaces.DimensionAbst;
+import com.kennycason.kumo.abst.DimensionAbst;
 
 import java.awt.*;
