+++ kumo-core/src/main/java/com/kennycason/kumo/awtimpl/FontImpl.java
@@ -1,6 +1,6 @@
 package com.kennycason.kumo.awtimpl;
 
-import com.kennycason.kumo.interfaces.FontAbst;
+import com.kennycason.kumo.abst.FontAbst;
 
 import java.awt.*;
 import java.io.IOException;
