+++ kumo-core/src/main/java/com/kennycason/kumo/collide/Collidable.java
@@ -1,8 +1,8 @@
 package com.kennycason.kumo.collide;
 
 import com.kennycason.kumo.image.CollisionRaster;
-import com.kennycason.kumo.interfaces.DimensionAbst;
-import com.kennycason.kumo.interfaces.PointAbst;
+import com.kennycason.kumo.abst.DimensionAbst;
+import com.kennycason.kumo.abst.PointAbst;
 
 /**
  * Created by kenny on 6/29/14.
