+++ kumo-core/src/main/java/com/kennycason/kumo/collide/RectanglePixelCollidable.java
@@ -2,8 +2,8 @@ package com.kennycason.kumo.collide;
 
 import com.kennycason.kumo.collide.checkers.RectanglePixelCollisionChecker;
 import com.kennycason.kumo.image.CollisionRaster;
-import com.kennycason.kumo.interfaces.DimensionAbst;
-import com.kennycason.kumo.interfaces.PointAbst;
+import com.kennycason.kumo.abst.DimensionAbst;
+import com.kennycason.kumo.abst.PointAbst;
 
 /**
  * Created by kenny on 7/2/14.
