+++ kumo-core/src/test/java/com/kennycason/kumo/examples/WordCloudNormalizersITest.java
@@ -5,8 +5,8 @@ import com.kennycason.kumo.WordCloud;
 import com.kennycason.kumo.WordFrequency;
 import com.kennycason.kumo.bg.CircleBackground;
 import com.kennycason.kumo.font.scale.LinearFontScalar;
-import com.kennycason.kumo.interfaces.DimensionAbst;
-import com.kennycason.kumo.interfaces.InstanceCreator;
+import com.kennycason.kumo.abst.DimensionAbst;
+import com.kennycason.kumo.abst.InstanceCreator;
 import com.kennycason.kumo.nlp.FrequencyAnalyzer;
 import com.kennycason.kumo.nlp.normalize.BubbleTextNormalizer;
 import com.kennycason.kumo.nlp.normalize.StringToHexNormalizer;
