+++ kumo-core/src/test/java/com/kennycason/kumo/palette/LinearGradientColorPaletteTest.java
@@ -1,10 +1,8 @@
 package com.kennycason.kumo.palette;
 
-import com.kennycason.kumo.interfaces.ColorAbst;
-import com.kennycason.kumo.interfaces.InstanceCreator;
+import com.kennycason.kumo.abst.ColorAbst;
+import com.kennycason.kumo.abst.InstanceCreator;
 import org.junit.Test;
-
-import java.awt.*;
 
 import static org.junit.Assert.assertEquals;
